# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "Laravel API",
        "contact": {
            "email": "QHs4o@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "config API Server"
        },
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User register",
                "description": "User register",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "gender"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "QHs4o@example.com"
                                    },
                                    "display_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "gender": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "message": "Register successfully",
                                                "data": {
                                                    "user": {
                                                        "id": 1,
                                                        "name": "John Doe",
                                                        "email": "QHs4o@example.com",
                                                        "display_name": "John Doe",
                                                        "gender": 1
                                                    },
                                                    "token": "xxxxxxxxx",
                                                    "type": "Bearer"
                                                }
                                            }
                                        },
                                        "error": {
                                            "type": "object",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null
                                        },
                                        "error": {
                                            "type": "object",
                                            "example": {
                                                "message": "Invalid data",
                                                "errors": {
                                                    "name": [
                                                        "The name field is required."
                                                    ],
                                                    "email": [
                                                        "The email field is required."
                                                    ],
                                                    "display_name": [
                                                        "The display name field is required."
                                                    ],
                                                    "password": [
                                                        "The password field is required."
                                                    ],
                                                    "gender": [
                                                        "The gender field is required."
                                                    ]
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Login",
                "operationId": "3c59992da251091e5a6c566d58c95396",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user",
                                    "password"
                                ],
                                "properties": {
                                    "user": {
                                        "oneOf": [
                                            {
                                                "description": "Email",
                                                "type": "string",
                                                "example": "QHs4o@example.com"
                                            },
                                            {
                                                "description": "Name",
                                                "type": "string",
                                                "example": "John"
                                            }
                                        ]
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "user": {
                                                    "id": 1,
                                                    "name": "John Doe",
                                                    "email": "QHs4o@example.com",
                                                    "display_name": "John Doe",
                                                    "gender": 1,
                                                    "active": 1,
                                                    "role": 0
                                                },
                                                "token": "xxxxxxxxx",
                                                "type": "Bearer"
                                            }
                                        },
                                        "error": {
                                            "type": "object",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null
                                        },
                                        "error": {
                                            "type": "object",
                                            "example": {
                                                "message": "Invalid data",
                                                "errors": {
                                                    "user": [
                                                        "The user field is required."
                                                    ],
                                                    "password": [
                                                        "The password field is required."
                                                    ]
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null
                                        },
                                        "error": {
                                            "type": "object",
                                            "example": {
                                                "message": "Login failed",
                                                "errors": {
                                                    "user": [
                                                        "Invalid username or password"
                                                    ]
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get myself information",
                "description": "Get myself information",
                "operationId": "getMe",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "message": "Get myself information successfully",
                                                "user": {
                                                    "id": 1,
                                                    "name": "John Doe",
                                                    "email": "QHs4o@example.com",
                                                    "display_name": "John Doe",
                                                    "gender": 1,
                                                    "active": 1,
                                                    "role": 0
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null
                                        },
                                        "error": {
                                            "type": "object",
                                            "example": {
                                                "message": "Invalid token"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}